refactor(web): move the submissions dashboard logic into domain/
pages/submissions/dashboard.ts was 1,753 lines of pure derivation (76
exports, zero React) consumed across three features, kept out of domain/
only by two type-only imports. The scores row types move from the
useGetScores hook to domain/submissions/scores.ts (the hook re-exports
them), BadgeTone comes from types/, and dashboard.ts moves beside them.
ActionListRow, a hub-modal row imported by both the submissions and
assignments features, moves to components/ui.
