+++ web/src/pages/AssignmentsPage.tsx
@@ -41,7 +41,7 @@ import useGetClassroomAssignments from "@/hooks/useGetClassAssignments"
 import useFunnelRoster, { unionLogins } from "@/hooks/useFunnelRoster"
 import { IncludeStaffToggle } from "@/pages/assignments/IncludeStaffToggle"
 import { persistIncludeStaff, readIncludeStaff } from "@/lib/includeStaffPref"
-import type { FunnelRoster } from "@/pages/submissions/dashboard"
+import type { FunnelRoster } from "@/domain/submissions/dashboard"
 import useGetClassroom from "@/hooks/useGetClassroom"
 import useEmptyRosterWarning from "@/hooks/useEmptyRosterWarning"
 import { useClassroomRoleContext } from "@/context/classroomRole/ClassroomRoleProvider"
