+++ web/src/pages/assignments/EditAssignmentForm.tsx
@@ -18,7 +18,7 @@ import { useTrackPublishDeploy } from "@/hooks/useTrackPublishDeploy"
 import { useEditAssignment } from "@/hooks/mutations/useEditAssignment"
 import { useAssignmentRepos } from "@/hooks/useAssignmentRepos"
 import useGetStudents from "@/hooks/useGetStudents"
-import { assignmentRepoNames } from "@/pages/submissions/dashboard"
+import { assignmentRepoNames } from "@/domain/submissions/dashboard"
 import { EditImpactConfirmModal } from "@/components/modals/EditImpactConfirmModal"
 import { LoadingSwap } from "@/lib/LoadingSwap"
 import { parseSubmissionTags } from "@/util/submissionTags"
