+++ web/src/pages/submissions/SubmissionsRowActions.tsx
@@ -16,10 +16,9 @@ import { createContext, useCallback, useContext, useState } from "react"
 import { Trans, useTranslation } from "react-i18next"
 
 import { safeHttpUrl } from "@/util/url"
-import { Button, EmphasisLtr } from "@/components/ui"
+import { ActionListRow, Button, EmphasisLtr } from "@/components/ui"
 import { ConfirmModal } from "@/components/modals"
 import { ActionIconLink } from "@/pages/submissions/SubmissionsRows"
-import { ActionListRow } from "@/pages/submissions/actionLayout"
 import { ReviewButton } from "@/pages/submissions/ReviewButton"
 import useTriggerRegrade from "@/hooks/useTriggerRegrade"
 import useDownloadSubmission from "@/hooks/mutations/useDownloadSubmission"
